requirements.txt
streamlit
cassandra-driver
langchain
sentence-transformers
openai
langchain-community